change to Long

+++ src/main/kotlin/com/hibob/academy/resource/OwnerResource.kt
@@ -1,16 +1,13 @@
 package com.hibob.academy.resource
 
-import jakarta.inject.Inject
 import jakarta.ws.rs.DELETE
 import jakarta.ws.rs.GET
 import jakarta.ws.rs.POST
 import jakarta.ws.rs.PUT
 import jakarta.ws.rs.Path
 import jakarta.ws.rs.PathParam
-import jakarta.ws.rs.Produces
 import org.springframework.stereotype.Controller
 import jakarta.ws.rs.core.Response
-import com.hibob.rsetApiDataClasses.Owner
 
 @Controller
 @Path("/api/owners")
@@ -20,25 +17,25 @@ class OwnerResource {
 
     @Path("/{ownerId}")
     @GET
-    fun getOwner(@PathParam("ownerId") ownerId: Int): Response {
+    fun getOwner(@PathParam("ownerId") ownerId: Long): Response {
         return Response.ok(ownerId).build()
     }
 
     @Path("/{ownerId}")
     @POST
-    fun postOwner(@PathParam("ownerId") ownerId: Int):Response {
+    fun postOwner(@PathParam("ownerId") ownerId: Long):Response {
         return Response.ok(ownerId).build()
     }
 
     @Path("/{ownerId}")
     @PUT
-    fun putOwner(@PathParam("ownerId") ownerId: Int):Response {
+    fun putOwner(@PathParam("ownerId") ownerId: Long):Response {
         return Response.ok(ownerId).build()
     }
 
     @Path("/{ownerId}")
     @DELETE
-    fun deleteOwner(@PathParam("ownerId") ownerId: Int): Response {
+    fun deleteOwner(@PathParam("ownerId") ownerId: Long): Response {
         return Response.ok(ownerId).build()
     }
 }
